Adds language switching functionality
Implements a language switch feature using the filament-language-switch package.

This allows users to change the application's language. It configures the available locales (ru, en, fr, he) and positions the switch in the top right corner of the admin panel.
Also moves socialite language files and adds the equivalent to the new supported languages.